implemented feedback from team, removed ffi directory in logins
