+++ megazords/ios/rust/src/lib.rs
@@ -8,7 +8,7 @@
 pub use crashtest;
 pub use fxa_client;
 pub use glean_ffi;
-//pub use logins_ffi;
+pub use logins;
 pub use nimbus;
 pub use places_ffi;
 pub use rc_log_ffi;
